+++ orm_lib/src/duckdb_impl/DuckdbConnection.h
@@ -1,7 +1,7 @@
 /**
  *
  *  @file DuckdbConnection.h
- *  @author An Tao
+ *  @author Dq Wei
  *
  *  Copyright 2018, An Tao.  All rights reserved.
  *  https://github.com/an-tao/drogon
@@ -56,11 +56,7 @@ class DuckdbConnection : public DbConnection,
                  std::function<void(const std::exception_ptr &)>
                      &&exceptCallback) override;
 
-    void batchSql(std::deque<std::shared_ptr<SqlCmd>> &&) override
-    {
-        LOG_FATAL << "DuckDB does not support batch mode yet";
-        exit(1);
-    }
+    void batchSql(std::deque<std::shared_ptr<SqlCmd>> &&sqlCommands) override;
 
     void disconnect() override;
 
@@ -84,6 +80,10 @@ class DuckdbConnection : public DbConnection,
     std::shared_ptr<DuckdbResultImpl> stmtExecute(
         duckdb_prepared_statement stmt);
 
+    // 批量 SQL 执行辅助方法 [dq 2025-11-21]
+    void executeBatchSql();
+    void executeSingleBatchCommand(const std::shared_ptr<SqlCmd> &cmd);
+
     trantor::EventLoopThread loopThread_;
 
     // DuckDB 数据库和连接句柄
@@ -100,6 +100,9 @@ class DuckdbConnection : public DbConnection,
 
     std::string connInfo_;
     std::unordered_map<std::string, std::string> configOptions_;  // DuckDB配置选项 [dq 2025-11-19]
+
+    // 批量 SQL 命令队列 [dq 2025-11-21]
+    std::deque<std::shared_ptr<SqlCmd>> batchSqlCommands_;
 };
 
 }  // namespace orm