+++ orm_lib/tests/db_test.cc
@@ -1,7 +1,7 @@
 /**
  *
  *  @file db_test.cc
- *  @author An Tao
+ *  @author Dq Wei
  *
  *  Copyright 2018, An Tao.  All rights reserved.
  *  Use of this source code is governed by a MIT license